Revert "start work on extracting mobile nav bar"
This reverts commit eb3473f28e70d6d65c9a62486d25f65f09f38f6d.
