+++ src/components/NavBar.svelte
@@ -59,7 +59,7 @@
 {/if}
 
 <style lang="scss">
-  @use "../../styles/base";
+  @use "../styles/base";
 
   .NavOpen {
     display: none;
